fix pytest issues on py36 and py27
