+++ lint.sh
@@ -2,9 +2,15 @@
 
 set -ex
 
+PY_VERSION=${1:-3}
+
 FWDIR="$(cd "`dirname $0`"; pwd)"
 cd "$FWDIR"
 
-prospector --profile "$FWDIR/prospector.yaml"
+if [[ "$PY_VERSION" -eq 3 ]]; then
+  prospector --profile "$FWDIR/prospector.yaml"
+else
+  prospector --profile "$FWDIR/prospector-2.yaml"
+fi
 
 rstcheck README.rst
