+++ prospector-2.yaml
@@ -1,0 +1,33 @@
+strictness: high
+test-warnings: True
+doc-warnings: false
+
+ignore-paths:
+  - sdk
+  - build
+  - env
+  - tunnel
+
+max-line-length: 100
+
+pylint:
+  disable:
+    - too-many-instance-attributes
+    - cyclic-import
+    - len-as-condition
+    - invalid-name
+    - no-else-return
+    - no-self-use
+    - protected-access
+    - too-many-arguments
+    - inconsistent-return-statements
+    - import-outside-toplevel
+    - consider-using-set-comprehension
+    - useless-object-inheritance
+    - unnecessary-pass
+    - superfluous-parens
+    - import-error
+
+mccabe:
+  disable:
+    - MC0001
