+++ tox-requirements-2.txt
@@ -1,8 +1,8 @@
 # Test reqs
-prospector[with_pyroma]==1.3.0
-pylint==2.5.2
+prospector[with_pyroma]==0.12.7
+pylint==1.8.2
 pep8-naming==0.5.0
-pytest==5.4.0
+pytest==3.8.1
 mock==2.0.0
 decorator==4.2.1
 rstcheck==3.2
@@ -11,5 +11,3 @@ pytest-cov==2.5.1
 docutils==0.14
 codecov
 requests_mock==1.5.2
-pytest-asyncio==0.16.0
-pytest-tornado==0.8.1
