+++ tox.ini
@@ -1,12 +1,15 @@
 [tox]
-envlist = py36
+envlist = py27, py36
 
 [testenv]
 # https://github.com/codecov/codecov-python/blob/5b9d539a6a09bc84501b381b563956295478651a/README.md#using-tox
 passenv = TOXENV CI TRAVIS TRAVIS_*
 deps =
-  py36: -rtox-requirements-3.txt
+  py27: -rtox-requirements-2.txt
+  py36: -rtox-requirements.txt
 commands =
-	pytest tests --cov=./
-	./lint.sh
+	py36: pytest tests --cov=./
+	py27: pytest tests --ignore=tests/tunnel --ignore="tests/sdk/test_command_api.py" --cov=./
+	py36: ./lint.sh
+	py27: ./lint.sh 2
 	codecov -e TOXENV
